+++ internal/pkg/node/constructors.go
@@ -6,6 +6,7 @@ import (
 	"syscall"
 
 	warewulfconf "github.com/warewulf/warewulf/internal/pkg/config"
+	"github.com/warewulf/warewulf/internal/pkg/hostlist"
 	"github.com/warewulf/warewulf/internal/pkg/wwlog"
 
 	"gopkg.in/yaml.v3"
@@ -25,7 +26,13 @@ func New() (NodesYaml, error) {
 	if err != nil {
 		return NodesYaml{}, err
 	}
-	return Parse(data)
+	nodeList, err := Parse(data)
+	if err != nil {
+		return nodeList, err
+	}
+	// add resolver for groups for subsequent calls
+	hostlist.SetGroupResolver(&nodeList)
+	return nodeList, nil
 }
 
 // Parse constructs a new nodeDb object from an input YAML